Fix #2477: Regression accessing `modules_to_save` (#2481)
* Fix #2477: Regression accessing `modules_to_save`

Commit ed3c82866ab3fb introduced adapter-local modules_to_save initialization which prevented
needless initialization but also broke prompt tuning methods as they don't have the `modules_to_save`
attribute.

This change also introduces a sequence classification test suite that also tests prompt tuning methods.
While not comprehensive it is sufficient to catch this error and can be extended over time.

While working on this and testing RoBERTa there was also an issue with the default target of `AdaLoRA`
as it defaults to `dense` (among other modules). This is problematic for `PeftModelForSequenceClassification`
as they mark `classification.*` as `modules_to_save`. But since the classification layer is also a dense layer
it will be targeted by `AdaLoRA`. To prevent such situations in the future a general excemption was made in
`check_target_module_exists` to always avoid keys in `modules_to_save`. For this to work the config modification
done in `PeftModelForSequenceClassification` needed changing.

* Remove presumably superflous code from inject_adapter

This code was *probably* for dealing with modules_to_save when calling
inject_adapter directly. However, since the only place that does this is
the PEFT mixed module which already deals with modules_to_save this
code is deemed superfluous.

This also makes dealing with ignoring `modules_to_save` in during targeting
easier since we can use the code in `check_target_module_exists` for every
case (targeting nested layer in modules_to_save module + direct targeting of
modules_to_save module).

* Move `set_additional_trainable_modules`

Move `set_additional_trainable_modules` to `inject_adapter` in case of adapters such as LoRAs, or,
in case of prompt tuning adapters, to their respective initialization point (while keeping the order
of operations intact).

Before this change a significant portion of `modules_to_save` initialization was removed from
`check_target_layer_exists` (called from `inject_adapter`) which only handled the `modules_to_save`
parameter in cases where this function was called directly (e.g., via `LoraModel.add_weighted_adapter`).
This also meant that trainable tokens was completely ignored in these cases. It also copied code from
`_set_trainable`.

The removal prompted the need to find a replacement which is this change: on adapter injection we will
now always check if there need to be additional trainable modules, not only during `PeftModel` init.